Date Uniformity + Missing Value Check
Converting the index to datetime, in case it was not doing that already. Following that, data is then sorted by date to ensure chronological order. Then, Checking for missing values and forward filling the data if anything is missing. After that, the data is dropped if still NaN after forward filling.